Construction duration estimation

+++ source/Stareater.Core/GameLogic/ConstructionResult.cs
@@ -4,12 +4,12 @@ namespace Stareater.GameLogic
 {
 	class ConstructionResult
 	{
-		public long DoneCount;
+		public double DoneCount;
 		public double InvestedPoints;
 		public Constructable Item;
 		public double PartialPoints;
 
-		public ConstructionResult(long donecount, double investedpoints, Constructable item, double partialpoints) {
+		public ConstructionResult(double donecount, double investedpoints, Constructable item, double partialpoints) {
 			this.DoneCount = donecount;
 			this.InvestedPoints = investedpoints;
 			this.Item = item;
